added settings container
